Fix space tab navigation
