+++ app/lib/features/space/pages/overview_page.dart
@@ -5,7 +5,6 @@ import 'package:acter/features/space/widgets/chats_card.dart';
 import 'package:acter/features/space/widgets/events_card.dart';
 import 'package:acter/features/space/widgets/links_card.dart';
 import 'package:acter/features/space/widgets/related_spaces_card.dart';
-import 'package:acter/features/space/widgets/top_nav.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
@@ -21,11 +20,6 @@ class SpaceOverview extends ConsumerWidget {
     // get platform of context.
     return Column(
       children: [
-        TopNavBar(
-          spaceId: spaceIdOrAlias,
-          key: Key('$spaceIdOrAlias::topnav'),
-          selectedKey: const Key('overview'),
-        ),
         Padding(
           padding: const EdgeInsets.all(20),
           child: StaggeredGrid.count(
