+++ app/lib/router/router.dart
@@ -30,11 +30,13 @@ import 'package:acter/features/settings/pages/labs_page.dart';
 import 'package:acter/features/settings/pages/licenses_page.dart';
 import 'package:acter/features/space/pages/overview_page.dart';
 import 'package:acter/features/space/pages/shell_page.dart';
+import 'package:acter/features/space/providers/space_navbar_provider.dart';
 import 'package:acter/features/todo/pages/create_task_sidesheet.dart';
 import 'package:acter/features/todo/pages/todo_page.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
+import 'package:riverpod/riverpod.dart';
 
 Future<String?> authGuardRedirect(
   BuildContext context,
@@ -73,369 +75,386 @@ final GlobalKey<NavigatorState> shellNavigatorKey =
 final GlobalKey<NavigatorState> spaceNavigatorKey =
     GlobalKey<NavigatorState>(debugLabel: 'space');
 
-final routes = [
-   GoRoute(
-    name: Routes.intro.name,
-    path: Routes.intro.route,
-    builder: (context, state) => const IntroPage(),
-  ),
-  GoRoute(
-    name: Routes.start.name,
-    path: Routes.start.route,
-    builder: (context, state) => const StartPage(),
-  ),
-   GoRoute(
-    name: Routes.introProfile.name,
-    path: Routes.introProfile.route,
-    builder: (context, state) => const IntroProfile(),
-  ),
-  GoRoute(
-    name: Routes.authLogin.name,
-    path: Routes.authLogin.route,
-    builder: (context, state) => const LoginPage(),
-  ),
-  GoRoute(
-    name: Routes.authRegister.name,
-    path: Routes.authRegister.route,
-    builder: (context, state) => const RegisterPage(),
-  ),
-  GoRoute(
-    path: '/gallery', 
-    builder: (context, state) => const GalleryPage(),
-  ),
-  GoRoute(
-    parentNavigatorKey: rootNavigatorKey,
-    name: Routes.bugReport.name,
-    path: Routes.bugReport.route,
-    pageBuilder: (context, state) => DialogPage(
-      builder: (BuildContext ctx) => BugReportPage(
-        imagePath: state.queryParameters['screenshot'],
+List<RouteBase> makeRoutes(Ref ref) => [
+      GoRoute(
+        name: Routes.intro.name,
+        path: Routes.intro.route,
+        builder: (context, state) => const IntroPage(),
       ),
-    ),
-  ),
-  GoRoute(
-    parentNavigatorKey: rootNavigatorKey,
-    name: Routes.quickJump.name,
-    path: Routes.quickJump.route,
-    pageBuilder: (context, state) => DialogPage(
-      builder: (BuildContext ctx) => const QuickjumpDialog(),
-    ),
-  ),
-  GoRoute(
-    parentNavigatorKey: rootNavigatorKey,
-    name: Routes.actionAddTask.name,
-    path: Routes.actionAddTask.route,
-    pageBuilder: (context, state) {
-      return SideSheetPage(
-        key: state.pageKey,
-        transitionsBuilder: (context, animation, secondaryAnimation, child) {
-          return SlideTransition(
-            position: Tween(
-              begin: const Offset(1, 0),
-              end: const Offset(0, 0),
-            ).animate(
-              animation,
-            ),
-            child: child,
-          );
-        },
-        child: const AddTaskActionSideSheet(),
-      );
-    },
-  ),
-  GoRoute(
-    parentNavigatorKey: rootNavigatorKey,
-    name: Routes.createSpace.name,
-    path: Routes.createSpace.route,
-    pageBuilder: (context, state) {
-      return SideSheetPage(
-        key: state.pageKey,
-        transitionsBuilder: (context, animation, secondaryAnimation, child) {
-          return SlideTransition(
-            position: Tween(
-              begin: const Offset(1, 0),
-              end: const Offset(0, 0),
-            ).animate(
-              animation,
-            ),
-            child: child,
-          );
-        },
-        child: CreateSpacePage(
-          initialParentsSpaceId: state.queryParameters['parentSpaceId'],
-        ),
-      );
-    },
-  ),
-
-  /// Application shell
-  ShellRoute(
-    navigatorKey: shellNavigatorKey,
-    // FIXME: unfortunately ShellRoute doesn't support redirects yet,
-    // thus we have to put it onto every route. Once that is fixed,
-    // remove that param from the sub-routes and use only here instead
-    // ref: https://github.com/flutter/flutter/issues/114559
-    // redirect: authGuardRedirect,
-
-    pageBuilder: (context, state, child) {
-      return NoTransitionPage(
-        key: state.pageKey,
-        child: HomeShell(child: child),
-      );
-    },
-    routes: <RouteBase>[
+      GoRoute(
+        name: Routes.start.name,
+        path: Routes.start.route,
+        builder: (context, state) => const StartPage(),
+      ),
+      GoRoute(
+        name: Routes.introProfile.name,
+        path: Routes.introProfile.route,
+        builder: (context, state) => const IntroProfile(),
+      ),
+      GoRoute(
+        name: Routes.authLogin.name,
+        path: Routes.authLogin.route,
+        builder: (context, state) => const LoginPage(),
+      ),
+      GoRoute(
+        name: Routes.authRegister.name,
+        path: Routes.authRegister.route,
+        builder: (context, state) => const RegisterPage(),
+      ),
       GoRoute(
         path: '/gallery',
         builder: (context, state) => const GalleryPage(),
       ),
       GoRoute(
-        name: Routes.myProfile.name,
-        path: Routes.myProfile.route,
-        redirect: authGuardRedirect,
+        parentNavigatorKey: rootNavigatorKey,
+        name: Routes.bugReport.name,
+        path: Routes.bugReport.route,
+        pageBuilder: (context, state) => DialogPage(
+          builder: (BuildContext ctx) => BugReportPage(
+            imagePath: state.queryParameters['screenshot'],
+          ),
+        ),
+      ),
+      GoRoute(
+        parentNavigatorKey: rootNavigatorKey,
+        name: Routes.quickJump.name,
+        path: Routes.quickJump.route,
+        pageBuilder: (context, state) => DialogPage(
+          builder: (BuildContext ctx) => const QuickjumpDialog(),
+        ),
+      ),
+      GoRoute(
+        parentNavigatorKey: rootNavigatorKey,
+        name: Routes.actionAddTask.name,
+        path: Routes.actionAddTask.route,
         pageBuilder: (context, state) {
-          return NoTransitionPage(
+          return SideSheetPage(
             key: state.pageKey,
-            child: const MyProfile(),
+            transitionsBuilder:
+                (context, animation, secondaryAnimation, child) {
+              return SlideTransition(
+                position: Tween(
+                  begin: const Offset(1, 0),
+                  end: const Offset(0, 0),
+                ).animate(
+                  animation,
+                ),
+                child: child,
+              );
+            },
+            child: const AddTaskActionSideSheet(),
           );
         },
       ),
       GoRoute(
-        name: Routes.activities.name,
-        path: Routes.activities.route,
-        redirect: authGuardRedirect,
+        parentNavigatorKey: rootNavigatorKey,
+        name: Routes.createSpace.name,
+        path: Routes.createSpace.route,
         pageBuilder: (context, state) {
-          return NoTransitionPage(
+          return SideSheetPage(
             key: state.pageKey,
-            child: const ActivitiesPage(),
+            transitionsBuilder:
+                (context, animation, secondaryAnimation, child) {
+              return SlideTransition(
+                position: Tween(
+                  begin: const Offset(1, 0),
+                  end: const Offset(0, 0),
+                ).animate(
+                  animation,
+                ),
+                child: child,
+              );
+            },
+            child: CreateSpacePage(
+              initialParentsSpaceId: state.queryParameters['parentSpaceId'],
+            ),
           );
         },
       ),
 
-      GoRoute(
-        name: Routes.tasks.name,
-        path: Routes.tasks.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
+      /// Application shell
+      ShellRoute(
+        navigatorKey: shellNavigatorKey,
+        // FIXME: unfortunately ShellRoute doesn't support redirects yet,
+        // thus we have to put it onto every route. Once that is fixed,
+        // remove that param from the sub-routes and use only here instead
+        // ref: https://github.com/flutter/flutter/issues/114559
+        // redirect: authGuardRedirect,
+
+        pageBuilder: (context, state, child) {
           return NoTransitionPage(
             key: state.pageKey,
-            child: const TodoPage(),
-          );
-        },
-      ),
-      GoRoute(
-        name: Routes.updates.name,
-        path: Routes.updates.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const NewsPage(),
+            child: HomeShell(child: child),
           );
         },
         routes: <RouteBase>[
-          // hide bottom nav for nested pages, use rootNavigatorKey
           GoRoute(
-            parentNavigatorKey: rootNavigatorKey,
-            name: Routes.updatesEdit.name,
-            path: Routes.updatesEdit.route,
+            path: '/gallery',
+            builder: (context, state) => const GalleryPage(),
+          ),
+          GoRoute(
+            name: Routes.myProfile.name,
+            path: Routes.myProfile.route,
             redirect: authGuardRedirect,
             pageBuilder: (context, state) {
               return NoTransitionPage(
                 key: state.pageKey,
-                child: const NewsBuilderPage(),
+                child: const MyProfile(),
               );
             },
           ),
           GoRoute(
-            parentNavigatorKey: rootNavigatorKey,
-            name: Routes.updatesPost.name,
-            path: Routes.updatesPost.route,
+            name: Routes.activities.name,
+            path: Routes.activities.route,
             redirect: authGuardRedirect,
             pageBuilder: (context, state) {
               return NoTransitionPage(
                 key: state.pageKey,
-                child: PostPage(
-                  attachmentUri: state.extra as String?,
+                child: const ActivitiesPage(),
+              );
+            },
+          ),
+
+          GoRoute(
+            name: Routes.tasks.name,
+            path: Routes.tasks.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const TodoPage(),
+              );
+            },
+          ),
+          GoRoute(
+            name: Routes.updates.name,
+            path: Routes.updates.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const NewsPage(),
+              );
+            },
+            routes: <RouteBase>[
+              // hide bottom nav for nested pages, use rootNavigatorKey
+              GoRoute(
+                parentNavigatorKey: rootNavigatorKey,
+                name: Routes.updatesEdit.name,
+                path: Routes.updatesEdit.route,
+                redirect: authGuardRedirect,
+                pageBuilder: (context, state) {
+                  return NoTransitionPage(
+                    key: state.pageKey,
+                    child: const NewsBuilderPage(),
+                  );
+                },
+              ),
+              GoRoute(
+                parentNavigatorKey: rootNavigatorKey,
+                name: Routes.updatesPost.name,
+                path: Routes.updatesPost.route,
+                redirect: authGuardRedirect,
+                pageBuilder: (context, state) {
+                  return NoTransitionPage(
+                    key: state.pageKey,
+                    child: PostPage(
+                      attachmentUri: state.extra as String?,
+                    ),
+                  );
+                },
+                routes: <RouteBase>[
+                  GoRoute(
+                    parentNavigatorKey: rootNavigatorKey,
+                    name: Routes.updatesPostSearch.name,
+                    path: Routes.updatesPostSearch.route,
+                    redirect: authGuardRedirect,
+                    pageBuilder: (context, state) {
+                      return NoTransitionPage(
+                        key: state.pageKey,
+                        child: const SearchSpacePage(),
+                      );
+                    },
+                  ),
+                ],
+              ),
+            ],
+          ),
+
+          GoRoute(
+            name: Routes.search.name,
+            path: Routes.search.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const SearchPage(),
+              );
+            },
+          ),
+          GoRoute(
+            name: Routes.chatroom.name,
+            path: Routes.chatroom.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const ChatPage(),
+              );
+            },
+          ),
+
+          GoRoute(
+            name: Routes.chat.name,
+            path: Routes.chat.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const ChatPage(),
+              );
+            },
+          ),
+
+          GoRoute(
+            name: Routes.dashboard.name,
+            path: Routes.dashboard.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const Dashboard(),
+              );
+            },
+          ),
+
+          // ---- SETTINGS
+
+          GoRoute(
+            name: Routes.info.name,
+            path: Routes.info.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const SettingsInfoPage(),
+              );
+            },
+          ),
+          GoRoute(
+            name: Routes.licenses.name,
+            path: Routes.licenses.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const SettingsLicensesPage(),
+              );
+            },
+          ),
+
+          GoRoute(
+            name: Routes.settings.name,
+            path: Routes.settings.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const SettingsMenuPage(),
+              );
+            },
+          ),
+
+          GoRoute(
+            name: Routes.settingsLabs.name,
+            path: Routes.settingsLabs.route,
+            redirect: authGuardRedirect,
+            pageBuilder: (context, state) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: const SettingsLabsPage(),
+              );
+            },
+          ),
+
+          /// Space subshell
+          ShellRoute(
+            navigatorKey: spaceNavigatorKey,
+            pageBuilder: (context, state, child) {
+              return NoTransitionPage(
+                key: state.pageKey,
+                child: SpaceShell(
+                  spaceIdOrAlias: state.pathParameters['spaceId']!,
+                  child: child,
                 ),
               );
             },
             routes: <RouteBase>[
               GoRoute(
-                parentNavigatorKey: rootNavigatorKey,
-                name: Routes.updatesPostSearch.name,
-                path: Routes.updatesPostSearch.route,
+                name: Routes.relatedSpaces.name,
+                path: Routes.relatedSpaces.route,
                 redirect: authGuardRedirect,
                 pageBuilder: (context, state) {
+                  debugPrint('in spaces related');
+                  ref
+                      .read(selectedTabKeyProvider.notifier)
+                      .switchTo(const Key('spaces'));
                   return NoTransitionPage(
                     key: state.pageKey,
-                    child: const SearchSpacePage(),
+                    child: RelatedSpacesPage(
+                      spaceIdOrAlias: state.pathParameters['spaceId']!,
+                    ),
+                  );
+                },
+              ),
+              GoRoute(
+                name: Routes.space.name,
+                path: Routes.space.route,
+                redirect: authGuardRedirect,
+                pageBuilder: (context, state) {
+                  debugPrint('in overview');
+                  ref
+                      .read(selectedTabKeyProvider.notifier)
+                      .switchTo(const Key('overview'));
+                  return NoTransitionPage(
+                    key: state.pageKey,
+                    child: SpaceOverview(
+                      spaceIdOrAlias: state.pathParameters['spaceId']!,
+                    ),
                   );
                 },
               ),
             ],
           ),
-        ],
-      ),
 
-      GoRoute(
-        name: Routes.search.name,
-        path: Routes.search.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const SearchPage(),
-          );
-        },
-      ),
-      GoRoute(
-        name: Routes.chatroom.name,
-        path: Routes.chatroom.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(key: state.pageKey, child: const ChatPage());
-        },
-      ),
-
-      GoRoute(
-        name: Routes.chat.name,
-        path: Routes.chat.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(key: state.pageKey, child: const ChatPage());
-        },
-      ),
-
-      GoRoute(
-        name: Routes.dashboard.name,
-        path: Routes.dashboard.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(key: state.pageKey, child: const Dashboard());
-        },
-      ),
-
-      // ---- SETTINGS
-
-      GoRoute(
-        name: Routes.info.name,
-        path: Routes.info.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const SettingsInfoPage(),
-          );
-        },
-      ),
-      GoRoute(
-        name: Routes.licenses.name,
-        path: Routes.licenses.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const SettingsLicensesPage(),
-          );
-        },
-      ),
-
-      GoRoute(
-        name: Routes.settings.name,
-        path: Routes.settings.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const SettingsMenuPage(),
-          );
-        },
-      ),
-
-      GoRoute(
-        name: Routes.settingsLabs.name,
-        path: Routes.settingsLabs.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const SettingsLabsPage(),
-          );
-        },
-      ),
-
-      /// Space subshell
-      ShellRoute(
-        navigatorKey: spaceNavigatorKey,
-        pageBuilder: (context, state, child) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: SpaceShell(
-              spaceIdOrAlias: state.pathParameters['spaceId']!,
-              child: child,
-            ),
-          );
-        },
-        routes: <RouteBase>[
           GoRoute(
-            name: Routes.relatedSpaces.name,
-            path: Routes.relatedSpaces.route,
+            name: Routes.spaces.name,
+            path: Routes.spaces.route,
             redirect: authGuardRedirect,
             pageBuilder: (context, state) {
-              debugPrint('in spaces related');
               return NoTransitionPage(
                 key: state.pageKey,
-                child: RelatedSpacesPage(
-                  spaceIdOrAlias: state.pathParameters['spaceId']!,
-                ),
+                child: const SpacesPage(),
               );
             },
           ),
+
           GoRoute(
-            name: Routes.space.name,
-            path: Routes.space.route,
-            redirect: authGuardRedirect,
-            pageBuilder: (context, state) {
-              debugPrint('in overview');
-              return NoTransitionPage(
-                key: state.pageKey,
-                child: SpaceOverview(
-                  spaceIdOrAlias: state.pathParameters['spaceId']!,
-                ),
-              );
+            name: Routes.main.name,
+            path: Routes.main.route,
+            redirect: (BuildContext context, GoRouterState state) async {
+              // we first check if there is a client available for us to use
+              final authGuarded = await authGuardRedirect(context, state);
+              if (authGuarded != null) {
+                return authGuarded;
+              }
+              if (isDesktop(context)) {
+                return Routes.dashboard.route;
+              } else {
+                return Routes.updates.route;
+              }
             },
           ),
         ],
       ),
-
-      GoRoute(
-        name: Routes.spaces.name,
-        path: Routes.spaces.route,
-        redirect: authGuardRedirect,
-        pageBuilder: (context, state) {
-          return NoTransitionPage(
-            key: state.pageKey,
-            child: const SpacesPage(),
-          );
-        },
-      ),
-
-      GoRoute(
-        name: Routes.main.name,
-        path: Routes.main.route,
-        redirect: (BuildContext context, GoRouterState state) async {
-          // we first check if there is a client available for us to use
-          final authGuarded = await authGuardRedirect(context, state);
-          if (authGuarded != null) {
-            return authGuarded;
-          }
-          if (isDesktop(context)) {
-            return Routes.dashboard.route;
-          } else {
-            return Routes.updates.route;
-          }
-        },
-      ),
-    ],
-  ),
-];
+    ];
